refactor(partners section): refactor partners section
refactor partners section, make hooks in separate files, change slider architecture
