+++ src/components/main/partners/Slider/Slider.tsx
@@ -1,0 +1,81 @@
+import React from 'react';
+import { RefObject, useRef } from 'react';
+import Image from 'next/image';
+import {
+  Swiper,
+  SwiperRef,
+  SwiperSlide,
+} from 'swiper/react';
+import useSwiperParams from './useSwiperParams';
+import useSliderControls from './useSliderControls';
+import { Navigation, Pagination } from 'swiper/modules';
+import Partner from '../partnerInterface';
+
+import 'swiper/css';
+import 'swiper/css/navigation';
+import 'swiper/css/pagination';
+import './styles.css';
+
+import ButtonRight from '@/components/icons/ButtonRight';
+import ButtonLeft from '@/components/icons/ButtonLeft';
+
+interface SliderProps {
+  partners: Partner[];
+}
+
+const Slider: React.FC<SliderProps> = ({ partners }) => {
+  const sliderRef: RefObject<SwiperRef> = useRef(null);
+  const swiperParams = useSwiperParams();
+  const { handlePrev, handleNext } =
+    useSliderControls(sliderRef);
+
+  return (
+    <div className="relative">
+      <Swiper
+        key={'partnersSlider pl-[100px]'}
+        modules={[Navigation, Pagination]}
+        slidesPerView={'auto'}
+        spaceBetween={swiperParams.spaceBetween}
+        navigation={{
+          prevEl: '.prev-partners',
+          nextEl: '.next-partners',
+        }}
+        pagination={{
+          clickable: true,
+        }}
+        loop={true}
+        className="xs:max-w-[280] sm:max-w-[380px] md:max-w-[618px] xl:max-w-[986px] 2xl:max-w-[1006px] 3xl:max-w-[1006px] 4xl:max-w-[1139px] 5xl:max-w-[1450px]"
+      >
+        {partners.map((partner) => (
+          <SwiperSlide
+            key={partner.id}
+            className="partnerSlide sm:max-w-[280] md:max-w-[190px] xl:max-w-[302px] 4xl:max-w-[324px] 5xl:max-w-[340px]"
+          >
+            <Image
+              src={partner.image}
+              width={swiperParams.imageWidth}
+              height={swiperParams.imageHeight}
+              alt={partner.alt}
+            />
+          </SwiperSlide>
+        ))}
+      </Swiper>
+      <button
+        type="button"
+        className="prev-partners absolute left-0 top-[24%] z-20 hidden translate-y-[-24%] cursor-pointer sm:hidden md:block"
+        onClick={handlePrev}
+      >
+        <ButtonLeft className="fill-white" />
+      </button>
+      <button
+        type="button"
+        className="next-partners absolute right-0 top-[24%] z-20 hidden translate-y-[-24%] cursor-pointer sm:hidden md:block"
+        onClick={handleNext}
+      >
+        <ButtonRight className="fill-white" />
+      </button>
+    </div>
+  );
+};
+
+export default Slider;
