+++ src/components/main/partners/Slider/styles.css
