+++ src/components/main/partners/Slider/useSliderControls.tsx
@@ -1,0 +1,20 @@
+import { RefObject, useCallback, useRef } from 'react';
+import { SwiperRef } from 'swiper/react';
+
+const useSliderControls = (
+  sliderRef: RefObject<SwiperRef>
+) => {
+  const handlePrev = useCallback(() => {
+    if (!sliderRef.current) return;
+    sliderRef.current.swiper.slidePrev();
+  }, [sliderRef]);
+
+  const handleNext = useCallback(() => {
+    if (!sliderRef.current) return;
+    sliderRef.current.swiper.slideNext();
+  }, [sliderRef]);
+
+  return { handlePrev, handleNext };
+};
+
+export default useSliderControls;
